function call support
Currently expects function to have an explicit
`~ return` statement. When the Ink function does
not have a return, the compiler seems to assume
that it will jump back after completing the
current container, so we need another way to
detect the end of the container and jump back to
the return element.