ink-proof integration tests
Runs the ink-proof test cases directly as Go tests.
Requires a clone of the ink-proof repo in `testdata/ink-proof`, as well as running inklecate to generate `story.ink.json` for each test.
This is currently managed with some local build scripts, but depends on an installation of inklecate, so hasn't been added to CI yet.