refactor tracking eval/string blocks
Using a counter for the depth doesn't work right with nested eval/str blocks. This keeps a reference back to the previous state each time we start to ensure we return to the expected state.
This also simplifies how we track implicit returns so that it's not scattered throughout the return statements for each evaluator.
add changes missing from refactor

These were sitting in a separate GitButler branch, need to merge them for the refactor.