tunnel choices preserve stack
Choices inside a tunnel should return to that stack frame when followed.
This passes ink-proof I077, I083, and I093.

Right now it returns to the stack as it was when the choice was created. This may need more testing for things like if variables were changed between the time the choice was created and following the choice.